requirements.txt
opencv-python==4.9.0.80
mss==9.0.1
pillow==10.1.0
pytesseract==0.3.10
numpy==1.24.4
requests==2.31.0
flask==3.0.0